Acc Sims!
added acceleration simulation outputs from the control system representing the longitudinal dynamics of a vehicle. This model was derived from the nonlinear system of equations detailed within this folder. It will serve as the nonlinear empiracle data to which LTI system will be matched and validated to for gain scheduling later in the project